requirements.txt
Django==5.1.2
psycopg[binary]==3.2.3
django-environ==0.11.2
celery==5.4.0
redis==5.0.8
requests==2.32.3
beautifulsoup4==4.12.3
newspaper3k==0.2.8
feedparser==6.0.11
jinja2==3.1.4
python-dateutil==2.9.0.post0
pydantic==2.9.2
structlog==24.1.0
sentry-sdk==2.17.0
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
freezegun==1.5.1
httpx==0.27.2
openai==1.51.2
numpy==2.1.2
